made some revisions
