+++ frontend/src/components/LandingPage/LandingPage.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Explanation from './Explanation.js/Explanation'
+import Explanation from './Explanation/Explanation'
 
 //Import components that you create here
